--- src/ibc_core/mod.rs
@@ -1,5 +1,0 @@
-pub mod channel;
-pub mod client;
-pub mod connection;
-pub mod port;
-pub mod routing;
